解决冲突,调试成功后正常运行

+++ boot-ssm/src/main/java/com/tequeno/bootssm/mapper/sys/UserPasswordMapper.java
@@ -6,5 +6,5 @@ import tk.mybatis.mapper.common.Mapper;
 
 public interface UserPasswordMapper extends Mapper<UserPassword> {
 
-    UserPassword selectByUserId(@Param("userId") Integer userId);
+    UserPassword selectByUserId(@Param("userId") Long userId);
 }
